feat: 토큰 스토어 제거 및 로컬스토리지 함수 추가
