+++ src/lib/localStorage.ts
@@ -1,0 +1,18 @@
+export const setToken = (token: string | null) => {
+  if (token) {
+    localStorage.setItem('token', token);
+  } else {
+    localStorage.removeItem('token');
+  }
+};
+
+export const getToken = (): string | null => {
+  if (typeof window !== 'undefined') {
+    return localStorage.getItem('token');
+  }
+  return null;
+};
+
+export const clearToken = () => {
+  localStorage.removeItem('token');
+};
