--- src/store/useAuthStore.ts
@@ -1,43 +1,0 @@
-'use client';
-
-import { useEffect } from 'react';
-import { create } from 'zustand';
-
-type AuthState = {
-  token: string | null;
-  setToken: (token: string | null) => void;
-  clearToken: () => void;
-};
-
-export const getToken = (): string | null => {
-  if (typeof window !== 'undefined') {
-    const token = localStorage.getItem('token');
-    return token;
-  }
-  return null;
-};
-
-export const useAuthStore = create<AuthState>((set) => ({
-  token: null,
-  setToken: (token) => {
-    if (token) {
-      localStorage.setItem('token', token);
-    } else {
-      localStorage.removeItem('token');
-    }
-    set({ token });
-  },
-  clearToken: () => {
-    localStorage.removeItem('token');
-    set({ token: null });
-  },
-}));
-
-export const useInitAuthStore = () => {
-  const setToken = useAuthStore((state) => state.setToken);
-
-  useEffect(() => {
-    const token = getToken();
-    if (token) setToken(token);
-  }, [setToken]);
-};
